+++ test/cypress/integration/example05.spec.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
 
-describe('Example 05 - Tree Data (from a flat dataset with parentId references)', () => {
+describe('Example 05 - Tree Data (from a flat dataset with parentId references)', { retries: 1 }, () => {
   const GRID_ROW_HEIGHT = 45;
   const titles = ['Title', 'Duration', '% Complete', 'Start', 'Finish', 'Effort Driven'];