+++ test/cypress/integration/example09.spec.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
 
-describe('Example 09 - OData Grid', () => {
+describe('Example 09 - OData Grid', { retries: 1 }, () => {
   beforeEach(() => {
     // create a console.log spy for later use
     cy.window().then((win) => {