+++ test/cypress/integration/example10.spec.js
@@ -8,7 +8,7 @@ function removeSpaces(textS) {
 const presetLowestDay = moment().add(-2, 'days').format('YYYY-MM-DD');
 const presetHighestDay = moment().add(20, 'days').format('YYYY-MM-DD');
 
-describe('Example 10 - GraphQL Grid', () => {
+describe('Example 10 - GraphQL Grid', { retries: 1 }, () => {
   it('should display Example title', () => {
     cy.visit(`${Cypress.config('baseExampleUrl')}/example10`);
     cy.get('h3').should('contain', 'Example 10 - Grid with GraphQL Backend Service');